Plugin and Swift EditorService Foundation
This patch starts the foundation of sending messages to Vim

The plugin logic will be implemented in Swift and interact with Vim
through sockets.

Currently, it uses HTTP for simplicity.

This approach provides isolation between the plugin and vim, and affords
easy async communication for the plugin that works across NeoVim and
Vim. This is a maintainable way to interface Vim <-> Swift.

Vim:

Bootstrap initial scaffolding.

EditorService:

Setup the core service to interact with the Vim UI.

Right now, it sends messages to the editor when the file .build/Last.log
changes. This is mainly a POC test and uses OSX specific APIs for now.