Show errors in .build/last_build.log on updates
This commit loads the error file. It sets a pretty generic error format,
and then loads the log into the quickfix window.

This makes it easy to see build warnings and errors in vim.

```
swift build ... | tee -a .build/last_build.log
```

Eventually, these should be signed to make it easir to see, similiar to
how iCompleteMe has done this.