Build Awareness UI Prototype Pass 1
This patch factors out diagnostic interface from iCompleteMe into a
simple python 2.7 program: the minimal requirements to display
diagnostics. It's fully compatible with iCompleteMe diagnostics and
produces a similar UI.

Currently, it finds errors from a raw build log in EditorService and
outputs models for the UI. The UI automatically updates and shows signs.

There are quite a bit of unideal and hacky parts to this. It is
mainly a prototype to get things working and experiment.

Ideally:
- Send an entire description of vim state to the UI via RPC
- Write a fast diagnostic UI and RPC layer without python