Swift Native Vim Bridge
This patch adds the ability to call vimscript from swift, and within the
Vim's main run loop.

Additonally, we an run vimscript from swift bycCall `command` and `exec`
through a high level interface "VimInterface"

VimCore: Vim -> Swift
- Vim calls our python helper program
- Calling a c function from python
- Calling `plugin_init` from the c function
VimCore is where `plugin_init` lives for SPMVim
This is the core bootstrap of the plugin

VimInterface: Swift -> Vim
- Call vim `eval` and command
- Swift functions interface with `swiftvim_*` API
- Internally, it uses Pythons C API

Currently, compile dylib with Makefile. It seems like SPM can handle
this task with some more effort, but that is left for a later exercise.

The fact that this uses Python as a glue layer is irrelevant. That
implementation detail should be hidden from Swift.

This is Step1 towards writing the entire plugin and a good RPC layer in
Swift.