Rewrite RPC Server in Swift
This patch ports most of the the RPC server from python into Swift. It
uses the POC http server for now. This is still mainly a prototype but
is working quite solid, so the old python version is gone.

There are still a few things left to be tidy'd up:
- consider adding a hook to tear down the server
- consider protobuf or msgpack for custom operations